requirements.txt
openai>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
pytest>=8.0.0
